+++ wallpapermgr/VERSION_HISTORY.rst
@@ -1,16 +1,30 @@
+Published to Pypy
+=================
 
 0.0.a1
-======
+------
 * First Release
 
 0.0.a2
-======
+------
 * Corrections to setup.py 
 
 0.0.a3
-======
+------
 * dependency: supercli updated, wallmgr needed update to maintain compatibility
 
 
+Unpublished to Pypy
+===================
+
+0.0.a4
+------
+* restructured class Archive() so that methods are used instead
+  or arguments to __init__... what was I thinking..
+
+* added ability to print all archives with descriptions
+
+TODO:
+   * ???